feat: Update rotated node types
Instead of assuming the node passed to the type can be rotated and
adding overloads for the different types of nodes, we can pass a generic
node to the rotated node types.

Now we don't need so many overloads on the rotation functions, as all of
the overloads would now be the same.